--- src/py21cmfast/drivers/composite_field.py
@@ -1,1 +1,0 @@
-"""Compute simulations that evolve over redshift."""
