+++ src/py21cmfast/utils.py
@@ -1,0 +1,35 @@
+"""Utilities for interacting with 21cmFAST data structures."""
+
+from .wrapper.outputs import InitialConditions, _OutputStructZ
+
+
+def get_all_fieldnames(
+    arrays_only=True, lightcone_only=False, as_dict=False
+) -> dict[str, str] | set[str]:
+    """Return all possible fieldnames in output structs.
+
+    Parameters
+    ----------
+    arrays_only : bool, optional
+        Whether to only return fields that are arrays.
+    lightcone_only : bool, optional
+        Whether to only return fields from classes that evolve with redshift.
+    as_dict : bool, optional
+        Whether to return results as a dictionary of ``quantity: class_name``.
+        Otherwise returns a set of quantities.
+    """
+    classes = [cls(redshift=0, dummy=True) for cls in _OutputStructZ._implementations()]
+
+    if not lightcone_only:
+        classes.append(InitialConditions())
+
+    attr = "pointer_fields" if arrays_only else "fieldnames"
+
+    if as_dict:
+        return {
+            name: cls.__class__.__name__
+            for cls in classes
+            for name in getattr(cls, attr)
+        }
+    else:
+        return {name for cls in classes for name in getattr(cls, attr)}
