+++ tests/test_data_exists.py
@@ -1,6 +1,6 @@
 import pytest
 
-from py21cmfast.inputs import DATA_PATH
+from py21cmfast.wrapper.inputs import DATA_PATH
 
 
 @pytest.fixture(scope="module")
